backend/requirements.txt
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-cors-headers==3.11.0
Pillow==9.0.1
openai==0.11.0
